Change contact detail on 7" to fragment carousel
Contact details in 7" landscape uses the fragment
carousel to show details with updates.  The layout
of the contact details fragment follows the 10"
landscape contact details fragment layout.  Dimming
of the detail or updates panel in the fragment
carousel has been disabled, and to be consistent,
this dimming is also disabled on the phone landscape
layout.

Also, this changes fixes the bug with the overscroll
indicators in the contact details pane.

Bug: 6398940
Bug: 6378743
Change-Id: Ie93ab73a9ebad8b66f5bc7d75222f5550ed1b005