Changed tablet tabs to be a drop down spinner
The tabs in the action bar have been removed for the
tablet form factors and replaced with the action bar
drop down navigation list.

Also, this changes the list item padding in the browse
list on the 7" tablet to extend a little further on the
border to the contact detail panel in portrait mode, so
that more text of longer names can be seen.

Bug: 6426927
Change-Id: Ieced44f5241c3f4f71bdbd7818846e24076fb355