Fix bug 6479405 On tablet, sometimes, QC -> full contact...
card yields in an empty details screen.

In the old code, we even copied the requested URI, which caused the issue
if the URIs for the first request and the second request are different
but still point at the same contact.  If we do this, the second result
will have a different requested URI than the actual requeseted URI,
which makes ContactLoaderFragment ignore the result.

This really happens because the lookup URI has variations.

Also make sure not to cache error results.

Bug 6479405

Change-Id: I71254578f0513b391804ee27a21a810bdb6881ac