Fixed contact detail animated transitions
When selecting different contacts on a tablet, the
animation is to set the contact details pane to white
and fade in the new contact details.

From quick contacts view (from Group or Favorites tab
and launcher screen shortcut), if quickcontact is opened
up and user touches the image, then when transitioning
back to the contact details, the contact details shows all
white and then transitions in the new contact details without
showing the previously selected contact details.

Bug: 6194409
Bug: 6324730
Change-Id: I5f56d46aef338b5d7a22e548548c42054656f381