Fix style of dropdown spinner on tablets
Changed text to be the same font size as the drop down
menu.  Capitalize only the first letter.  Changed the
color of the spinner triangle to the dark action bar
color.  And changed the on-press highlight color to match
that of the other action bar items.

Also, took out all exit fade animations, and fixed the
presence icon size on the contact detail list for the
7" tablet.

Bug: 6500339
Change-Id: Ia6c5555e1ad772812672c37d5b722e2a5aaa40ed