Tweak quick contact startup
Start the loader before inflating the layout.

Layout inflation is a CPU bound operation + file system
access if the layout isn't on the filesystem cache.

This CL increases the chance (unfortunately it's not 100%) for
the layout inflation and the data load to overlap.

Also added "null stopwatch" to easily disable stopwatch.

Bug 6434853

Change-Id: I8be2d68bec3997c967633abe7726be7ff4a61c89