Fix pop-up menus for photos in contact editor
The hasMoreThanOnePhoto() has been re-implemented to
look directly at the entity deltas to determine whether
there are more than one photo in the set of raw contacts
rather than relying on whether a photo was set in a raw
contact editor view.

Also, calls to refresh the raw contact views are made if
any changes to raw contact photos are made.

Finally, fixed a regression where the photo handler is not
set correctly so that the wrong photo handler would handle
the callback onPhotoSelected which caused a newly taken
image or chosen image from the gallery would get set to the
wrong raw contact.

Bug: 6408705
Bug: 5850397
Change-Id: I94e3872ed3e0191f1ca845de0d01525aee3a5f5f