Fix blank edit view after orientation change.
If an orientation change occurs before the contact data loads from the
database, the code fails to reload the data because it is checking
against saved state (which is always present on an orientation change).
This fix changes the check to look for the actual data to be present
in the saved state.

Bug: 7122054
Change-Id: I4373eca14ba447abc57493cf4c5af6b978d8b7a2