Fix bug 7250807 Ellipsizing is broken
The ellipsis wasn't showing in names in the Contacts app. It turns out
that the problem was simply a discrepancy between onMeasure and onLayout
for the TextView displaying the name: the former wasn't taking into
account the mTextIndent. Thus, the actual TextView was short 8 dip,
resulting in truncation on the right (which usually resulted in loss of
the ellipsis).

This change simply takes mTextIndent into account when doing the
measurement of the child TextView for mNameTextView.

Change-Id: I750d32c6444937d87a7dfc25850ba7e486b3c6ae