Implement lift-to-type for phone dialer accessibility.
When explore by touch is turned on, lifting a finger inside of a dialer
key will perform that key's action. This code also ensures that calling
performClick() on a key performs the key's action (req'd for Braille).

During touch exploration, keys are set to be non-clickable so that lift-
to-type doesn't conflict with normal double-tap interaction.

Bug: 7263951
Change-Id: I7d4a292ebc306b51666d9aa840ddcb2a7e5770fb