Update People app to direct SMS SEND_TO only to the default SMS app.
In KLP we have added the concept of a default SMS app to the platform. Only
this application has write permission to the SMS database. To help ensure
a better user experience we will always direct "send to sms" in the people
app to the app that the user has selected as their default SMS app.

Bug: 10870624 System should direct "SENDTO" intent with "sms" and "mms" schemes to the default SMS app
Change-Id: Ic35282b52d0f12baa11b57dc1b3ae8556ecb9a5a