Adjust layout for dialpad to support both Dialer and InCallUI.
This mostly was done by changing the key rows so that their height
is determined by layout weights instead of explictly defined. The
size of some objects were adjusted as well, to better fit the grids
for the newer design.

Changed text dimensions to be dp instead of sp, so they don't scale.

Bug: 14471388
Change-Id: Ie29376bd485f12970f117d2bc97e47733fd92717