Implement custom view for rendering numerals.
This is a custom text view intended only for rendering the numerals
(and star and pound) on the dialpad. TextView has built in
top/bottom padding to help account for ascenders/descenders.

Since vertical space is at a premium on the dialpad, particularly
if the font size is scaled to a larger default, for the dialpad we
use this class to more precisely render characters according to the
precise amount of space they need.

Change-Id: I0eb18b2b2c1f31fa87aae30a71f51b323a9906a3