Set first header's y position according to the to of the first child.
We're applying padding in Dialer, but this is not being picked up in
the ListView because the pinned header was absolutely positioned and
did not respect padding. In this addition, if the first child of the
list is visible, check it's top and set the first header to have
this value if the top is non-negative.

Bug: 15332945
Change-Id: Ia3b46cb6bd69547faff57ae8f94b48b598adb670