Fix bad text kerning in dynamic resizing EditText.
This used to work with setTextScaleX, I thought, but now it seems
that changing that property only affects the character and not the
spacing, which looks sad.

This fixes by scaling text size instead. This is a slightly indirect
method of doing this, because text size doing this not map directly
to width, but this appears to be effective nonetheless. It scales
more smoothly with better spacing; I can fit up to ~19 digits now.

Bug: 15457260
Change-Id: I93bef566766246a86000c6658d058cc1688adc46