Add support for full-bleed selection of contact list items.
- Adding the concept of "adjustSelectionBoundsEnabled", which allows
us to control whether the headerWithoutBounds is used to override the
selection list items bounds. We want to be able to do this in some
cases (eg. in the contacts list when there is a header on the left)
but in others (eg. a search list), we want the ripple to fill out the
entire bounds of the list item.
- Allow for this to be set by the ContactEntryListFragment, and then
passed through the ContactEntryListAdapter to the ContactListItemView.
- No longer apply scrollbar padding to the list view regardless of
whether the scrollbar is present, because this messes with the ripple
propagation.

Bug: 15517457
Change-Id: I174a462ac1e45b987ac0ee70de318ed037d91762