Add in-app CountryDetector (1/4)
Replace use of the hidden CountryDetector system service with one that
runs in-app.

Instead of using a LocationManager listener to listen to location updates,
the in-app CountryDetector uses pending intents to passively listen to
any location updates.

This change also introduces an IntentService that is used to simplify the
asynchronous geocoding of the location that is provided to the CountryDetector's
BroadcastReceiver by the LocationManager

Bug: 15593973

Change-Id: I2915aad176963a04c7c577addb659984f0db56f2