In Contacts, QC image should start full size
This means that QC won't initially show contact photos as tinted
when opening them inside Contacts.

We still show QC contact photos as tinted when opening them outside
Contacts. This is because we need to perform cropping when opening
QC outside Contacts. And, cropped photos look bad without a tinting.

Bug: 15892293
Change-Id: Ic34303d32a40c9ea5f88c8a8ebd88ce1ea8e261d