Update avatar drawable in LetterTileDrawable
Use the same image as QC uses. Except, instead of scaling down
the ~140dp image start with a 64dp image.

The images provided by allen don't have an amount of border built into
it that looks good inside the LetterTileDrawable. Therefore, I wrapped
the image inside a ScaleDrawable. To do this, I changed LetterTileDrawable
to operate on internal drawables instead of internal bitmaps.

Bug: 16290899
Change-Id: I741c3a603d025068614c7004ea61fb349f6d6944