Increase max QC title size on 720dp+ tablets
Because of the way that View#setScaleY() works, the maximum size of the
title TextView has to be less than the minimum size of the of the TextView's
parent. As a result, increasing the title size caused unexpected behavior.

Fixing this odd behavior required me to remove the title TextView
out of the header ViewGroup and instead manually control its position
inside MultiShrinkScroller. This required me to make MultiShrinkScroller
a FrameLayout. This necessitated lots of little changes inside MultiShrinkScroller
(ie, what type of LayoutParams to cast into).

There is a reason nobody else has messed around with scaling title TextViews.

Bug: 16683381
Change-Id: Ib8fc065931a6a0c2a96b8bd5a11bb055457aa3dd