Update color of FAB
I copied Dialers fab into Contacts instead of ContactsCommon.

Since this is a visual element, there is little chance of it
diverging from Dialer in an undetected way.

I was previously overzelously trying to unify colors/icons between
dialer and contacts. This is not a good policy when others are less
zeoulos.

Bug: 16847357
Change-Id: I6e3587d5871a485963f8677e9697ee777d6f1caf