Use semi-stable ids in Contact lists
I manually tested the postal, contact, email and phone pickers.

Note that the legacy pickers & email picker don't refresh
when the dataset changes, so this doesn't have any affect on them.
Nonetheless, I set viewIds for these pickers in case we ever change
these pickers to refresh when the underlying dataset changes.

This is Part 2/2 for b/17195707. This fixes contact-pickers.

Bug: 17195707
Change-Id: I6e72768ab9f30f11d632983794da1ed36346416c