Fix scroll offset calculation for collapsed header
When originally writing the logic that handled flinging and EdgeEffect
it was impossible for contacts viewed inside the Contacts app
to have a collapsed contact photo without the contact photo
touching the top of the window. Now that we have added click-to-expand
and click-to-collapse to the contact photo, we need to handle one
more case.

Bug: 17405547
Change-Id: I0fbad31b96ae7cf182708e9386b5327dfda8a226