Avoid using donottranslate_config
I should have avoided using donottranslate_config.xml. I assume
this file works the same as donottranslate.xml. But I'm not certain
and I don't want to risk this. Plus, its a weird file which is used
for manual translations and help link overlays.
Mixing it with non-translated strings is going to confuse people.

Bug: 17668051
Change-Id: I395e081ed96af11a51cf18c9d9d07f2ef20bb612