Keep landscape quickcontact title on screen
When QuickContacts opens partially offset from the top of
the screen, the title should be visible. To do this, I
offset the title padding by the amount that QuickContacts is
offset from the top of the screen, up to a maximum amount
equal to the starting QuickContact offset.

See bug for demo mp4s.

Bug: 17417615
Change-Id: Ie39c130c23c9541bdd912bbb406a9bae0966e965