New design for photo editor
The photo editor field now looks like the mocks. It includes
a RadioButton that allows users to choose which photo is
primary. This RadioButton only appears if more than one raw
contact has a photo inside the contact.

Additional changes:
-fixed non-regression that made it impossible to
 set some photos as primary
-fixed non-regression that made it possible for multiple
 photos to appear as primary
-I fudged with a bunch of padding values in order to
 compensate for View insets and paddings. This can't always
 be done perfectly, since some insets/paddings vary based
 on screen-configuration. I needed to start taking fudging
 with these paddings in order to make the photo editor look
 consistent with the other editor fields.

This CL doesn't bother ensuring the readonly photo edit
field looks good. This will be done in a later CL.

Bug: 18004959
Change-Id: I683f3ffa8e7f8683c2c083a2b4592b2b95c58c63