Use buildDataStringForDisplay for phone field
This just fixes the format of the number as displayed
inside the phone field.

It doesn't fix the formatting inside recent interactions:
we continue to display the exact phone number returned
from the sms and call log database. When the contact
doesn't have a name, we continue to show their name
as an unformatted number.

Bug: 7230016
Change-Id: I1d8b40c27fbd3497a4fef4e1f30ab2a21eca44d4