Handle reading loaded Contact data in the base edit fragment
This reduces the number of abstract methods sub classes
must implement and also keeps us from having to duplicate
the logic to set up the new/local/me/existing RawContactData
List (mState) in all sub edit fragments.

The crux of this CL is moving the "bindEditor" methods
into the base fragment.

Also renamed the various bindEditors methods which are
mostly reading fields the loaded Contact and settting up
mState to be "setState" since we it seemed more appropriate
to reserve the name bindEditors for setting up the actual
Views.

Bug 19124091

Change-Id: Ieb4a906ac372faa20694e546caa3bcba8c6b53b4