Exit selection mode after batch delete or merge
We still don't support this for batch shares. It difficult
to know whether a share intent was received or whether the
disambiguation dialog was cancelled. In order to do this,
you need to use a timer to guess whether the app has been
backgrounded or not. Since sharing is a non-destructive
action, I think it is almost appropriate to stay in selection
mode after sharing anyway.

Bug: 19549465
Change-Id: Ibb132504e7a19ba2866de84bca216d126e4fbfaa