Layout mExpansionViewContainer when not displayed
Force mExpansionViewContainer to always be laid out. This will ensure
that the whitespace on the end of mExpansionViewContainer is correct
regardless of whether an expansion view is needed or not.

This fixes the way phonetic name is displayed on exchange contacts.
This has no affect on focus contacts since both expansion views
are set to View.VISIBLE for focus contacts.

Bug: 19503535
Change-Id: I28aad309158a21a07d38a9699f82b48d72b8e7c4