Stop using static common test lib. Fix Jack build
In order to depend on TestCommon inside this test, we can
do either of the following
1) Use a static library
2) Directly use TestCommon code

Option 1 requires us to manually recompile the static library
(or use mma) every time we want to change the static library.
Even worse, because of changes to Jack we aren't guaranteed
that the static library will be built the first time for us.
If we wanted to, we could fix this by directly executing
TestCommon's makefile inside this makefile.

Option 2, essentially makes TestCommon code a part of
this project. It gets recompiled every time you run mm.

Bug: 19667686
Change-Id: Ib31c5cfda8a7d7e55e63b2d308307574ea1e2612