Add 24dp person icon to ContactsCommon.
Surely this icon is already added somewhere, but apparently not
in ContactsCommon. Since it's a pretty "core" icon, I figure it
makes more sense to add it here than in Dialer, where I'm going
to be referencing it, since I figure it is likely it will be
referenced or should be referenced by other places in the future.

Bug: 19372539
Change-Id: Ib1571974638d7c835f780ed9a767e58a5f34158a