1/2 Allow the keyboard to resize the compact editor layout
So we're reverting the fix submitted in ag/680795 and
preventing the landscape layout from getting messed up
by basing the photo resizing off of the display dimensions.

The reason is that adjustResize provides a better UX and
is necessary to fix the bugs related to showing and hiding
the type downdowns propertly.

Bug 20526470

Change-Id: Ia6755336cb8d1fa78142b754e5d40a45b4c50d44