Distinguish between editor back button presses and framework stopage
We carry whether the editor fragment is being stopped
because of a back button press or because the framework
stopped the hosting Activity all the way through the various
editor fragment callbacks and the ContactSaveService because
it is not until ContactEditorBaseActivity.onSaveFinished
where we start the next Intent -- starting it causes a
"flash" if recents is clicked and follwed by an immediate
starting of the next editor Activity, which is the bug
that was filed.

With this change, we only use the ContactSaveService
resultIntent to go back to the compact editor on back
presses.  The expected behavior/tested scenarios are
described at go/b21198041

Bug 21198041
Bug 19624360

Change-Id: Ic350e12aa447cff81747e003f504f25100bd5c60