Fix detection of empty structured names
TextFieldEditorView.isEmpty does not do the right thing
for structured names because that methods just checks every
child of the fields ViewGroup and the view retains values
for fields that may have been erased by toggling the
name expansion.  The result is that the FIELD_TURNED_EMPTY
EditorListener event is never fired.

Bug 21851290

Change-Id: Iaae3ba102a899de9dfd76d0469c9c2f1d094b0bd