Ignore new raw contact display names when saving contacts
Now that we have logic to ignore the display name
we add to newly created raw contacts (when the
user wants to edit a read-only contact) when
determining whether the user made any edits,
we can enable changing the display on the compact
editor again.

Note, there is one glitch -- the name disappears
on the compact editor after you click more fields
(but only when editing read-only contacts).

Tested scenarios:

1) Edit read-only contact, more fields, back, back

   No bogus contact is created and the read-only
   name is displayed on both editors.

2) Edit read-only contact, change name on compact
   editor, back, re-edit read-only contact, and split

   The read-only contact is unchanged and a new
   contact with the edited name is created.

3) Edit read-only contact, more fields, change name on
   full editor, back, back, re-edit read-only contact,
   and split

   Same result as (2)

4) Edit read-only contact, add phone on compact
   editor, back, re-edit read-only contact, and split

   The read-only contact is unchanged and a new
   contact with the read-only name and phone number is
   created.

5) Edit read-only contact, more fields, add phone on full
   editor, back, back, re-edit read-only contact, and split

   Same result as (4)

Bug 21858251
Bug 21464081

Change-Id: I9028fee38b8ea7569968654a756eb98025318e69