Remove isVisibility check before scaleIn.
Turns out, assumption that we'd never want to scaleIn a visible FAB is false.

If the FAB is in the middle of a transition (eg. scaling out), it cancels
that animation, and some states can conflict. It's possible to fix these in
this statement, but... at this point, it just seems more conservative just
to leave it the way it was working before and address the FAB case in Dialer
differently.

Bug: 21938194
Bug: 22070516
Change-Id: I5e5f88af71faed4ce603e21b00f526e829e6d3fc