Fix showing all phone nums and emails for joined contacts
We must delete the entire KindSectionView for emails and
phone numbers for all but the last raw contact that is being
edited.

Also, icons must be updated whenever an editor housed in
a kindview is deleted since we only want to show the icon
once for all the kindviews added to the compact layout
for all the raw contacts in the delta list.

Bug 22203548

Change-Id: If12c4fc10a9c9c09cbcfb86c5bf1472d89f6b890