Allow the user to select the vcards to import
Rather than always scanning for .vcf files in external storage

This is the flip-side to b/22453182 where we changed to allow
the user to select the destination for vcard export so that
we could remove the WRITE_EXTERNAL_STORAGE permission.

Additionally, we change the import/export options in the
dialog to read "Export to .vcf file" and "Import from .vcf file"
since the previous text (Just "Export" and "Import from external
storage" are not accurate.

Note, we continue to support importing from multiple .vcf files
at once but we no longer need our own dialog to present those
options to the user.

Bug 22454038

Change-Id: I1a145f00f07c4c0f547c38d720c109197baab833