Apply contact join ops before we reach the max batch op threshold
Splitting the join operations across several batches "could" lead
to some parts of the join being applied and the remaining parts
not applied because of some error.

Our new UX to join more than two contacts also increases the
possibility that the joins take a long time.  This is especially
true when a contact involved in the joined is composed of many
raw contacts since the current contact save logic creates an
operation between every pair of raw contacts.

Bug 22480225

Change-Id: Icab836f587808e018bd59909864dfd9de0d53776