Consolidate CallSubject history and dialog.
- Removed CallSubjectHistory activity and put that functionality into
CallSubjectDialog.
- Changed CallSubjectDialog into an activity styled after a dialog.
- Although dialog_call_subject looks a lot different, the main change is
the addition of the subject_list above the dialog in the layout (I tried
having it below the dialog but I could not get the layout to work
properly).  A RelativeLayout was used so that when the subject history
list is shown, the dialog naturally moves up to fit the call history list.
- Added layout for the call subject history list items to give better
control of their appearance.
- This was to give better control of layout and make it possible to
animate between show/hide of call subject history.


Bug: 22685114
Change-Id: I7555bb09b86c50b52a76661f1460f4fe0e30bd97