Revert moving between the compact and full editors (E5)
We are reverting most of the changes made to fix
b/19624360 in MNC.  The reason to do this is to reduce
the massive complexity moving between the two editors
introduced.

Since the compact editor now expands in place, the full
editor will likely only used to view read-only raw
contacts now.

Bug 23589603

Change-Id: Idfcc07d8aae65295d4b3f0a0e3e7f5eb701a46c4