Show all aggreagate contact names on the editor
Stop copying the name from read only contacts
to the new raw contact created in order to enable
editing of read only contacts. We also don't need to
match the name displayed by quick contacts.
In the new design, we are showing multiple names with
read-only accounts marked with the account type icon
and name.

Bug 23589603

Change-Id: Ifd8d18b17fc9645c15814bbb1a649a82fe246bbd