Move finding primary/writable raw contacts to KindSectionDataList (E16)
This will help us since editing the 1) name and 2) photo, and 3) finding
the primary account used to sort editors so that new fields get added
to the primary account all do something similar to find the
ValuesDelta that changes are written too.

We can also enforce that only the same DataKind/mime-type are added
to a KindSectionDataList and passed to a KindSectionView.

Bug 24509375
Bug 23589603

Change-Id: I2fc603c3cea2fb7f10a96c8592688ccf3deef56a