Show check mark when none is primary in photo picker
If there is a photo in photo picker that is marked as primary,
we will use check that photo in photo picker.

Otherwise, with this CL, we compare the photo URI passed from
QuickContact to editor with the photo URIs in photo picker to
decide where to add a check mark.

Bug: 25092609
Change-Id: I0ae77f9874a60f45dcfb305d058ade9de4fef4a4