Backport setter methods
The following methids were added in API level 21 and we have to ignore
them on older SDKs:
1. setOutlineProvider
2. setTranslationZ
3. setElegantTextHeight
4. setOverlay

View.setTranslationZ could be replaced with ViewCompat.setTranslationZ,
but that method basically ignores the attribute on older SDKs, which is
equivalent to our approach here.

In addition, we backport setImageTintList by applying tint to a drawable
and setting the drawble to ImageView.

Bug: 25629359
Change-Id: I3c5f7fe972cc0a5f7c8ab67581b0e68c3cdad35b