Call forceLoad() when loader is not started to load hi-res photo
The bug occurs because the loader may not actually be in the state
of "started" after calling restartLoader. If this happens, QuickContact
will not get the newly updated hi-res photo since onContentChanged()
will not call forceLoad(). So we call forceLoad() explicitly in this
case. There will be exactly 1 extra load when we call forceLoad()
and will be no extra load otherwise.

Bug 25204200

Change-Id: I41cc35381f7b24c0d461a4b8ed482b675e8ccad9