Removing final from ContactsPreferences
As a final class, ContactsPreferences was not mockable. This placed
undue complexity on testing code because rather than mocking the
ContactsPreferences, mocking the dependencies of this object was
required.

Since this code is not part of a library, removing the final modifier
is the quickest way to more easily testable code. If we need
ContactsPreferences to be final, we need to create an interface for
ContactsPreferences, implement it with a concrete class and implement
a concrete MockContactsPreferences object so we can test code that
uses this object.

Change-Id: I276436b6f7db0f87e3432f3fafe1d177bc6c5e50