Changing unit test case to verify account selection dialog is disabled.
Previously, whether an account is default or not
relied on the setting in preferences.

CL 820506 changed the logic to, when trying to get default account,
look at the writable account queue first. If it's size is exactly 1,
directly return that account as default, without looking into preferences.

Updating matching unit test case to adapt to this change.

Change-Id: I170b7cb2f37c5fdeedb6e23bc7eab53b63d7a1e7