ContactsCommon: Fix duplicate strings
This CL fixes duplicate string resource errors I got when migrating
ContactsCommon from make to gradle.

It looks like the root cause of the error is the use of
product="tablet" and product="nosdcard". As far as I can tell this
approach for resources only works for bundled apps.

Here were the duplicate resources and their resolution:
  - account_phone - moved to values-sw600dp/strings.xml
  - fail_reason_too_many_vcard - deleted, used in ExportVCardActivity
    but that code was also unused
  - composer_has_no_exportable_contact - moved to
    values-sw600dp/strings.xml
  - import_failure_no_vcard_file - deleted, unused
  - scanning_sdcard_failed_message - deleted, unused

Bug: 26676586
Change-Id: I2a95c269215bee381ad95aa2f4fbb28c2d43d8fa