Bring back code to share favorite contacts
ag/865209 led to an error that pressing "share favorite contacts"
would share all contacts. For favorite and frequently contacted
contacts, we should still use the previous method to fetch and
share, because there will not be many of them and it won't take
long.

In addition, we also show toast in UI thread

Bug 22083005
Bug 26906732
Bug 26964977

Change-Id: I4db6bc4d79cec5ec75197accb3ac049e0fe29722