Fix crash in ContactSelectionActivity during adding widget
The crash was introduced when we included appcompat library.

custom_action_bar.xml is only used in ContactSelectionActivity which is
not appcompat yet; so we should use the regular SearchView like before.

search_menu.xml is also only used in the activity and should be reverted
back too.

Bug 28074607
Bug 27687799

Change-Id: Iaad4753939a14d71c08639a5e73aae1e703bcd27