Show icon of account being saved to
Remove unused layouts for various account header types and simplify to
two:
One for linked contacts (editor_all_rawcontacts_account_selector) which
displays a linked chain icon and one which displays the account icon and
an optional expander icon (editor_account_header). The account icon one
is used in the single account case and when creating a new contact.

Test:
Manual usage of the new selector:
1. Single account case
2. Linked account case shows linked chain with a selector dropdown
3. Creating a new contact shows icon of account being saved to
3b. 3 but with multiple accounts so the selector is visible 
3c. 3b with swapping accounts and making sure the icon updates correctly

Bug: 29455526
Change-Id: Ic7784cc6e77b934468795176ee774a5247cce9a9