Focus label edit view on click
Take focus away from previously focused fields so that the
the background view doesn't scroll up when the new label
dialog is shown.

Test:
Manual
Tested opening the create new label dialog while another
edit text had focus.
Tested canceling the dialog and committing a new label.
In the case of committing made sure the label edit view
was in the viewport so the entire popup is visible.

Bug: 29055820
Bug: 29211998

Change-Id: I45c604f9b94ab03a67bc1d5c64d98364140f931f