Editor expander changes
Missed tinting the phonetic name expander in ag/1417505.
Went ahead and did some refactoring since
phonetic_name_edit_expansion_view was the same as the name_edit one
but with a different content description. Removed it since content
description should be dynamic with the state of the expander anyway.
Along that line also made accessibility improvements and fixed a bug
with the state of the icon after rotation.

Test:
Checked the color of the phonetic name expander.
Actioned the expanders with Talkback on and made sure the correct
announcements were made.
Checked state of icon after rotation was consistent.

Bug: 30160325
Bug: 31033553
Change-Id: Ia62ed1a03b4edda81815387e3bcf4f48449699e2