Remove multi account grouping of data in compact editor
Remove most List versions of KindSectionData and KindSectionViews,
with exception of photo section data, since we still want to show
the correct photo. In the aggregate screen with the multi account
selector, we show the photo of the aggregate in read only. Once
a raw contact is chosen to edit, we show the photo of that raw
contact.
Removed sorting that dealt with multiple accounts. Still sorting
the mimetypes.
Drop custom field data kinds until we're ready to render them.
Use for loops instead of for each where possible.

Test: Tested the following editor scenarios:
  1) new contact
  2) edit other contact
  3) edit writable raw contact
  4) edit read-only raw contact (joins a new writable raw contact to it)
  5) edit aggregate w/ 1 writable and 1 read-only raw contact
  6) edit aggregate w/ 2 writable raw contacts
  7) edit local me raw contact
  8) edit local me raw contact joined with a read-only raw contact
  9) editing photo of a raw contact that is part of an aggregate
     updates the aggregate photo
  10) same with editing the name

Bug: 31088704
Change-Id: I39219af0bfd1ac0899b6518983be942d78bda9aa