Restrict aggregation suggestions
Only show contacts which have a raw contact that's in the same
account as the raw contact that's currently being edited.
This also implies no single read only contact will be suggested
since you can't trigger the suggestions from a read-only account.

Test:
Tested entering a name that matches another contact within the same
account shows the suggestion popup when:
  1) suggested contact has only 1 writable raw contact.
  2) suggested contact has multiple writable raw contacts.
  3) suggested contact has a mix of writable and read only contacts.
Checked no suggested contact appears when entering a contact name
that exists in a different account
Checked the above is still true when switching between the different
accounts in the account switcher.

Bug:31704115
Bug:31088704
Change-Id: I45e36e1e61b04069af419c47669c338985eb583d