Show sync-off alert
In all contacts list:
alert user when global sync is off if user never dismissed before,
until sync setting is toggled on then off again.

In account view:
when global sync is off, show alert with the same logic for all
contacts list;
when global sync is on, alert user when contacts sync for this account
is off if user never dismissed before, until contacts sync is toggled
on then off again.

Bug: 31755559

Test: Manually tested the scenarios:
1) Add/remove account,
2) switch between all contacts list, account view, and label views,
3) update contacts information,
3) turn on sync/dismiss from alert,
when global sync is on/off, contacts syncs are all on/off, contacts
sync is on/off for a particular account.

Change-Id: Id5001847270d5a984942c63a42edb5f9401bc2c2