Implement editor springboard activity
Have the springboard handle edit intents.
Show a dialog if the contact is made of multiple raw contacts.
Go straight to editor if:
  * Single raw contact
  * Given a raw contact Uri
In the case of 1 read 1 writable, we still show the dialog
since there would otherwise be no way to view what data comes
from the read only raw contact.
The springboard does not handle legacy contact Uris and will throw
an exception if one is received.

Test:
Tested these scenarios:
 1) Edit a single raw contact
 2) Edit a single read only raw contact
 3) Edit a contact made of >2 raws
 4) Add new contact
 5) Edit a contact made of one read only, one writable
 6) Made edits and checked if quick contact continued to update
 7) The relevant edit intents from the Test app

Bug: 31826229
Bug: 31088704
Change-Id: I4c1c44accc86521efce2081744189d25f00ec541