Reintroduce pseudo mimetype for name
In order to have a separate edit DataKind for phonetic name
we need a pseudo mimetype for just the name as well. The pseudo
DataKinds are used to render the actual text fields while the
StructuredName mimetype DataKind is used for checking if any
thing was changed.

Test:
Unit tests
Save a new contact while entering only the phonetic name

Bug: 32023197
Bug: 21758689

Change-Id: I151b37b7be091705974e3f36770e4bb0e150a64d