Use given DefaultImageRequest when loading from Uris
Add empty bytes array check. Without this the photo would be
blank if one did not exist.
The given DefaultImageRequest was only used if the Uri was null,
changed Request to also hold the given DIR to use if a photo couldn't
be loaded.

Test:
Manual usage of the raw contact picker dialog.

Bug: 31826229
Bug: 32096517
Change-Id: Ied938552326f8a258b558040e1b1c1b5d5723239